+++ ChiTech/ChiMath/SpatialDiscretization/FiniteElement/PiecewiseLinear/pwlc_03_buildsparsity.cc
@@ -321,8 +321,8 @@ BuildSparsityPattern(std::vector<int64_t> &nodal_nnz_in_diag,
   nodal_nnz_in_diag.clear();
   nodal_nnz_off_diag.clear();
 
-  nodal_nnz_in_diag.resize(local_base_block_size*N,0);
-  nodal_nnz_off_diag.resize(local_base_block_size*N,0);
+  nodal_nnz_in_diag.resize((size_t) local_base_block_size*N,0);
+  nodal_nnz_off_diag.resize((size_t) local_base_block_size*N,0);
 
   if (unknown_manager.dof_storage_type == chi_math::UnknownStorageType::NODAL)
   {
